+++ fresh.gen.ts
@@ -7,8 +7,9 @@ import * as $1 from "./routes/api/joke.ts";
 import * as $2 from "./routes/index.tsx";
 import * as $$0 from "./islands/AboutSection.tsx";
 import * as $$1 from "./islands/ContactList.tsx";
-import * as $$2 from "./islands/Counter.tsx";
-import * as $$3 from "./islands/ProjectList.tsx";
+import * as $$2 from "./islands/Container.tsx";
+import * as $$3 from "./islands/Counter.tsx";
+import * as $$4 from "./islands/ProjectList.tsx";
 
 const manifest = {
   routes: {
@@ -19,8 +20,9 @@ const manifest = {
   islands: {
     "./islands/AboutSection.tsx": $$0,
     "./islands/ContactList.tsx": $$1,
-    "./islands/Counter.tsx": $$2,
-    "./islands/ProjectList.tsx": $$3,
+    "./islands/Container.tsx": $$2,
+    "./islands/Counter.tsx": $$3,
+    "./islands/ProjectList.tsx": $$4,
   },
   baseUrl: import.meta.url,
 };
